requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-ollama>=0.1.0
langchain-core>=0.3.0
pydantic>=2.0.0
streamlit>=1.38.0
python-dotenv>=1.0.0
requests>=2.31.0
duckduckgo-search>=3.9.0
tavily-python>=0.3.0